Move constants ENV_VARS, DEFAULT_PROXY_HOSTNAME, DEFAULT_PROXY_PORT to apify-shared-js, move name of stores emulations dirs to apify-shared-js
