+++ package.json
@@ -45,7 +45,7 @@
   },
   "dependencies": {
     "apify-client": "^0.2.7",
-    "apify-shared": "^0.0.29",
+    "apify-shared": "^0.0.31",
     "bluebird": "^3.5.0",
     "content-type": "^1.0.3",
     "fs-extra": "^5.0.0",
